refactor: 클래스 등록 관련 mutation form 훅 내부로 수정
