+++ src/pages/Admin/Class/hooks/useClassRegisterForm.ts
@@ -1,10 +1,9 @@
-import { useQueryClient } from '@tanstack/react-query';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { isAxiosError } from 'axios';
 import { SubmitHandler, useForm } from 'react-hook-form';
 
+import AdminClassApi from '@/services/admin/class';
 import { FieldRules } from '@/types';
-
-import useClass from './useClassMutation';
 
 interface IClassRegisterForm {
   title: string;
@@ -48,7 +47,14 @@ function useClassRegisterForm({ onClose }: UseClassRegisterFormProps) {
   };
 
   const queryClient = useQueryClient();
-  const { classMutation } = useClass();
+
+  const classMutation = useMutation({
+    mutationFn: (classData: {
+      title: string;
+      type: string;
+      description: string;
+    }) => AdminClassApi.createClass(classData),
+  });
 
   const submitForm: SubmitHandler<IClassRegisterForm> = (data, event) => {
     event?.preventDefault();
