+++ src/pages/Admin/Class/hooks/useSubClassRegisterForm.ts
@@ -1,10 +1,9 @@
-import { useQueryClient } from '@tanstack/react-query';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { isAxiosError } from 'axios';
 import { SubmitHandler, useForm } from 'react-hook-form';
 
+import AdminClassApi from '@/services/admin/class';
 import { FieldRules } from '@/types';
-
-import useClass from './useClassMutation';
 
 interface ISubClassRegisterForm {
   classId: number;
@@ -48,7 +47,15 @@ function useSubClassRegisterForm({ onClose }: UseSubClassRegisterFormProps) {
   };
 
   const queryClient = useQueryClient();
-  const { subClassMutation } = useClass();
+
+  const subClassMutation = useMutation({
+    mutationFn: (subClassData: {
+      classId: number;
+      title: string;
+      description: string;
+      orderNumber: number;
+    }) => AdminClassApi.createSubClass(subClassData),
+  });
 
   const submitForm: SubmitHandler<ISubClassRegisterForm> = (data, event) => {
     event?.preventDefault();
